update code fix error upload file

+++ src/main/java/com/nhom44/validator/NumberVallidator.java
@@ -5,7 +5,7 @@ public class NumberVallidator implements SingleValidator {
     public boolean validator(String input) {
         if (input==null||input.trim().isEmpty()) return false;
         try {
-            int num = Integer.parseInt(input);
+            double num = Double.parseDouble(input);
             return true;
         } catch (NumberFormatException e) {
             return false;
